none/sw.js: Remove functioning fetch handler.
This is for consistency with all the other apps (there is no point
having only the 'none' app having a proper service worker). This service
worker is currently too annoying to use for development because its
cache number needs to be manually updated each time, so I am simply
removing the caching functionality for now. It will be brought back in
some form in the future.